+++ Example/ManaKit/AppDelegate.swift
@@ -34,11 +34,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 //        DatabaseMaintainer.sharedInstance.rules2CoreData()
 
          // Temporary updates
-//         DatabaseMaintainer.sharedInstance.updateArtist()
+//        DatabaseMaintainer.sharedInstance.updateArtist()
+//        DatabaseMaintainer.sharedInstance.updateCardIDs()
         
           // Normal run
         ManaKit.sharedInstance.setupResources()
-        ManaKit.sharedInstance.configureTCGPlayer(partnerKey: "ManaGuide", publicKey: nil, privateKey: nil)
+        ManaKit.sharedInstance.configureTCGPlayer(partnerKey: "ManaGuide", publicKey: "A49D81FB-5A76-4634-9152-E1FB5A657720", privateKey: nil)
         
         return true
     }